Made listening address configurable (#38)
The client disallows http (no 's') by default but the server only
serves http. The server is therefore likely to be configured
behind a revers proxy. A defense-in-depth approach involves
having such services listen only on localhost. This change makes
the address the server listens on configurable to support such a
configuration.