Move info function from package main into goolib. (#170)
The info() function in googet.go is used by the "installed" and "available" subcommands to print out info about a PkgSpec. This moves that code into the goolib package, next to the PkgSpec definition. Cleaned it up, made it a method on PkgSpec, and renamed it to PrettyPrint. Added tests. The only difference in the old output vs the new output is that listed dependencies are now printed in a deterministic order.